@@ -1,6 +1,6 @@
 {
   "name": "ngx-select-ex",
-  "version": "3.5.13-dev",
+  "version": "3.6.1",
   "description": "Angular based replacement for select boxes",
   "license": "MIT",
   "private": false,
@@ -26,7 +26,7 @@
     "publish.ng4": "npm publish dist --tag ng4",
     "publish.ng5": "npm publish dist",
     "publish.ng5-dev": "npm publish dist --tag dev",
-    "start": "ng serve -op runtime",
+    "start": "ng serve",
     "test": "ng test --watch false",
     "test-coverage": "ng test --watch false --code-coverage"
   },
@@ -42,6 +42,7 @@
     "angular2",
     "angular4",
     "angular5",
+    "angular6",
     "select",
     "select2",
     "ui-select",
@@ -54,50 +55,52 @@
     "@angular/platform-browser": ">5.0.0"
   },
   "devDependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cli": "1.6.7",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/animations": "^6.1.0",
+    "@angular/cli": "~6.2.3",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
     "@types/marked": "^0.3.0",
-    "@types/node": "~6.0.60",
+    "@types/node": "~8.9.4",
     "@types/webpack": "^3.8.4",
     "codecov": "^3.0.0",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "~4.3.0",
     "commitizen": "^2.9.6",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.4",
     "cz-conventional-changelog": "^2.1.0",
     "escape-string-regexp": "^1.0.5",
     "html-loader": "^0.5.5",
     "husky": "^0.14.3",
-    "jasmine-core": "~2.8.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "lodash": "^4.17.4",
     "markdown-loader": "^2.0.2",
-    "ng-packagr": "^2.0.0",
-    "ngx-bootstrap": "^2.0.2",
+    "ng-packagr": "^4.2.0",
+    "ngx-bootstrap": "^3.0.1",
     "pre-commit": "^1.2.2",
-    "protractor": "~5.1.2",
-    "rxjs": "^5.5.6",
+    "protractor": "~5.4.0",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "standard-version": "^4.3.0",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "zone.js": "^0.8.19"
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~2.9.2",
+    "zone.js": "~0.8.26"
   },
   "config": {
     "commitizen": {
